requirements.txt
streamlit
boto3
pdf2image
pytesseract
opencv-python
Pillow
